requirements.txt
Pillow
tqdm
opencv-python
